refactor: extract AIProvider enum from SettingsStore.swift (#635)
Extracts the AIProvider enum (119 lines, 4 cases + 6 computed properties)
from Sources/BugNarrator/Services/SettingsStore.swift into its own file.

Byte-identical move; SHA-256 eb2f9626fea4e74e0bd7689771af3db0c21ee6d27a378dd38050d432678f3c61
verified against origin/main:SettingsStore.swift lines 4-122.

Imports: Foundation (for URL type).

Reduces SettingsStore.swift from 2,083 lines to 1,964 lines. Self-contained
enum extraction — no callers reference by file location.

Tests: 667 (unchanged).
